requirements.txt
torch~=1.7.0
torchvision~=0.8.1
setuptools~=50.3.1
numpy~=1.19.2
h5py~=3.3.0
Pillow~=8.0.0
natsort~=7.1.0